+++ SDLGameEngine/ObjectLayer.cpp
@@ -62,7 +62,7 @@ void ObjectLayer::render(){
                 
                 //If light is meant to scroll, then offset it
                 if (m_gameObjects[i]->GetParams().isScrolling())  {
-                    Vector2D<float> pos(cameraParams.getX(),
+                    Vector2D<double> pos(cameraParams.getX(),
                                  cameraParams.getY());
                     
                     cameraParams.setX(pos.getX());
